Add name validation and clean up code

+++ src/main/java/view/PublisherListBean.java
@@ -4,11 +4,14 @@ import bridge.PublisherBridge;
 import entity.Publisher;
 import jakarta.annotation.PostConstruct;
 import jakarta.faces.application.FacesMessage;
+import jakarta.faces.component.UIComponent;
 import jakarta.faces.context.FacesContext;
+import jakarta.faces.validator.ValidatorException;
 import jakarta.faces.view.ViewScoped;
 import jakarta.inject.Named;
 import lombok.Getter;
 import lombok.Setter;
+import org.jetbrains.annotations.NotNull;
 import org.primefaces.PrimeFaces;
 
 import java.io.Serializable;
@@ -46,38 +49,38 @@ public class PublisherListBean implements Serializable {
         PrimeFaces.current().ajax().update("dialogs");
     }
 
-
-    public void deletePublishers() {
-        List<Publisher> toDeletePublisher = new ArrayList<>();
-        FacesContext context = FacesContext.getCurrentInstance();
-
-
-        for(Publisher publisher : selectedPublishers){
-            List<Object[]> MagazinesWithPublisherNames = publisherBridge.getMagazinesWithPublisherNames(publisher);
-            if (MagazinesWithPublisherNames.isEmpty()) {
-                toDeletePublisher.add(publisher);
-            }
-            else{
-                for (Object[] result : MagazinesWithPublisherNames) {
-                    String magazineTitle = (String) result[0];
-                    String publisherName = (String) result[1];
-                    FacesMessage message = new FacesMessage(FacesMessage.SEVERITY_ERROR,
-                            "Löschen fehlgeschlagen",
-                            "Magazine " + magazineTitle + " hat Verlag " + publisherName + " noch als Link!");
-                    context.addMessage(null, message);
-                }
-            }
+    public void deletionFailed(@NotNull List<Object[]> MagazinesWithPublisherNames){
+        for (Object[] result : MagazinesWithPublisherNames) {
+            String magazineTitle = (String) result[0];
+            String publisherName = (String) result[1];
+            FacesContext.getCurrentInstance().addMessage(null, new FacesMessage(FacesMessage.SEVERITY_ERROR,
+                    "Löschen fehlgeschlagen",
+                    "Magazine " + magazineTitle + " hat Verlag " + publisherName + " noch als Link!"));
         }
-        publisherBridge.deleteMultiplePublishers(toDeletePublisher);
-        updateAllPublisher();
-        FacesMessage message = new FacesMessage(FacesMessage.SEVERITY_INFO,
-                "Löschen erfolgreich",
-                "Es wurden " + toDeletePublisher.size() + " von " + selectedPublishers.size() + " Verlagen gelöscht!");
-        context.addMessage(null, message);
-        PrimeFaces.current().ajax().update("form:messages", "form:allPublisherList");
     }
 
 
+    public void deletePublishers() {
+        List<Publisher> toDeletePublisher = new ArrayList<>();
+
+        for (Publisher publisher : selectedPublishers) {
+            List<Object[]> MagazinesWithPublisherNames = publisherBridge.getMagazinesWithPublisherNames(publisher);
+            if (MagazinesWithPublisherNames.isEmpty()) {
+                toDeletePublisher.add(publisher);
+            } else {
+                deletionFailed(MagazinesWithPublisherNames);
+            }
+        }
+
+        if(!toDeletePublisher.isEmpty()){
+            publisherBridge.deleteMultiplePublishers(toDeletePublisher);
+            updateAllPublisher();
+            FacesContext.getCurrentInstance().addMessage(null,new FacesMessage(FacesMessage.SEVERITY_INFO,
+                    "Löschen erfolgreich",
+                    "Es wurden " + toDeletePublisher.size() + " von " + selectedPublishers.size() + " Verlagen gelöscht!"));
+        }
+        PrimeFaces.current().ajax().update("form:messages", "form:allPublisherList");
+    }
 
 
     public void deletePublisher(Publisher publisher) {
@@ -88,20 +91,12 @@ public class PublisherListBean implements Serializable {
             updateAllPublisher();
             FacesMessage message = new FacesMessage(FacesMessage.SEVERITY_INFO,
                     "Löschen erfolgreich",
-                    "Der Verlag "+publisher.getName()+ " wurde erfolgreich gelöscht.");
+                    "Der Verlag " + publisher.getName() + " wurde erfolgreich gelöscht.");
             context.addMessage(null, message);
-            PrimeFaces.current().ajax().update("form:messages", "form:allPublisherList");
         } else {
-            for (Object[] result : MagazinesWithPublisherNames) {
-                String magazineTitle = (String) result[0];
-                String publisherName = (String) result[1];
-                FacesMessage message = new FacesMessage(FacesMessage.SEVERITY_ERROR,
-                        "Löschen fehlgeschlagen",
-                        "Magazine " + magazineTitle + " hat Verlag " + publisherName + " noch als Link!");
-                context.addMessage(null, message);
-                PrimeFaces.current().ajax().update("form:messages");
-            }
+            deletionFailed(MagazinesWithPublisherNames);
         }
+        PrimeFaces.current().ajax().update("form:messages", "form:allPublisherList");
     }
 
     public boolean hasSelectedPublishers() {
@@ -127,5 +122,13 @@ public class PublisherListBean implements Serializable {
         }
         updateAllPublisher();
         PrimeFaces.current().ajax().update("form:allPublisherList");
+    }
+
+    public void nameValidation (FacesContext context, UIComponent component, Object value) throws ValidatorException {
+        if(publisherBridge.isPublisherNameTaken((String) value)){
+            throw new ValidatorException(new FacesMessage(FacesMessage.SEVERITY_ERROR,
+                    "Name vergeben!",
+                    "Der Name" + value + " is bereits vergeben."));
+        }
     }
 }
